Refactoring de string calculator en otra clase.
